python-backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.5.3
python-dotenv>=1.0.0
httpx>=0.26.0
pydantic-settings>=2.0.0

# LLM Providers
google-generativeai>=0.3.0
langfuse>=2.0.0

# Audio Processing & Transcription
assemblyai>=0.25.0
supabase>=2.3.0
tenacity>=8.2.0
webrtcvad>=2.0.10  # Voice Activity Detection for user identification
aiofiles>=23.0.0  # Async file operations for streaming downloads

# Audio Processing
numpy>=1.24.0,<2.0.0  # Used by VAD and audio processing
soundfile>=0.12.1  # Used by audio processing

# Image Processing
Pillow>=10.0.0  # Used for pre-scaling images in video export

# Email
resend>=0.8.0

# Error Tracking & Monitoring
sentry-sdk[fastapi]>=2.35.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
pytest-mock>=3.12.0

